requirements.txt
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pycodestyle==2.5.0
PyJWT==1.7.1
requests==2.22.0
urllib3==1.25.8
